lib-common/hash: use ssize_t for len of hash functions.
The hash functions imported from platform with 99815f9b113, needs to
have the same arguments as the iop hash functions.

With 6c709c9cf64, we now use ssize_t for the len argument for the iop
hash functions.
So, we need to do the same modifications as 9fabe2747a9f, and use
ssize_t for the len argument of the functions imported with 99815f9b113.

+++ hash.h
@@ -149,16 +149,16 @@ static inline uint32_t u64_hash32(uint64_t u64)
     return (uint32_t)(u64) ^ (uint32_t)(u64 >> 32);
 }
 
-uint64_t identity_hash_64(const void * nonnull data, int len);
+uint64_t identity_hash_64(const void * nonnull data, ssize_t len);
 
-uint64_t murmur3_128_hash_64(const void * nonnull data, int len);
+uint64_t murmur3_128_hash_64(const void * nonnull data, ssize_t len);
 
-static inline uint64_t crc64_hash_64(const void * nonnull data, int len)
+static inline uint64_t crc64_hash_64(const void * nonnull data, ssize_t len)
 {
     return icrc64(0, data, len);
 }
 
-static inline uint64_t hsieh_hash_64(const void * nonnull data, int len)
+static inline uint64_t hsieh_hash_64(const void * nonnull data, ssize_t len)
 {
     return hsieh_hash(data, len);
 }
